+++ chb/ast/ASTApplicationInterface.py
@@ -39,7 +39,7 @@ from chb.ast.CustomASTSupport import CustomASTSupport
 from chb.ast.ASTNode import ASTStmt, ASTVarInfo
 
 
-pirversion: str = "0.1.0-20220923"
+pirversion: str = "0.1.0-20220929"
 
 
 class ASTApplicationInterface: